+++ noodles-vcf/src/variant/record/samples/series/value/array/values.rs
@@ -1,4 +1,6 @@
-use std::{io, iter};
+use std::{borrow::Cow, io, iter};
+
+use crate::io::reader::record_buf::value::percent_decode;
 
 /// Variant record sample array value values.
 #[allow(clippy::len_without_is_empty)]
@@ -81,23 +83,23 @@ impl<'a> Values<'a, char> for &'a str {
     }
 }
 
-impl<'a> Values<'a, &'a str> for &'a str {
+impl<'a> Values<'a, Cow<'a, str>> for &'a str {
     fn len(&self) -> usize {
         count(self)
     }
 
-    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<Option<&'a str>>> + '_> {
+    fn iter(&self) -> Box<dyn Iterator<Item = io::Result<Option<Cow<'a, str>>>> + '_> {
         if self.is_empty() {
             Box::new(iter::empty())
         } else {
-            Box::new(
-                self.split(DELIMITER)
-                    .map(|s| match s {
-                        MISSING => None,
-                        _ => Some(s),
-                    })
-                    .map(Ok),
-            )
+            Box::new(self.split(DELIMITER).map(|s| {
+                match s {
+                    MISSING => Ok(None),
+                    _ => percent_decode(s)
+                        .map(Some)
+                        .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)),
+                }
+            }))
         }
     }
 }
@@ -120,15 +122,21 @@ mod tests {
     #[test]
     fn test_string_values() -> io::Result<()> {
         let src = "";
-        let values: Box<dyn Values<'_, &'_ str>> = Box::new(src);
+        let values: Box<dyn Values<'_, Cow<'_, str>>> = Box::new(src);
         assert_eq!(values.len(), 0);
 
         let mut iter = values.iter();
         assert!(iter.next().transpose()?.is_none());
 
-        let src = "a,b,.";
-        let values: Box<dyn Values<'_, &'_ str>> = Box::new(src);
+        let src = "a,b%3Ac,.";
+        let values: Box<dyn Values<'_, Cow<'_, str>>> = Box::new(src);
         assert_eq!(values.len(), 3);
+
+        let mut iter = values.iter();
+        assert_eq!(iter.next().transpose()?, Some(Some(Cow::from("a"))));
+        assert_eq!(iter.next().transpose()?, Some(Some(Cow::from("b:c"))));
+        assert_eq!(iter.next().transpose()?, Some(None));
+        assert!(iter.next().transpose()?.is_none());
 
         Ok(())
     }